unattached-pd/requirements.txt
#Copyright 2019 Google LLC
#Licensed under the Apache License, Version 2.0 (the "License");
#you may not use this file except in compliance with the License.
#You may obtain a copy of the License at
#https://www.apache.org/licenses/LICENSE-2.0
#Unless required by applicable law or agreed to in writing, software
#distributed under the License is distributed on an "AS IS" BASIS,
#WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
#See the License for the specific language governing permissions and
#limitations under the License.

Flask==0.12.2
Flask-WTF==0.14.2
google==2.0.1
google-api-core==1.1.0
google-api-python-client==1.7.7
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-cloud-core==0.28.1
google-cloud-monitoring==0.28.1
googleapis-common-protos==1.5.3
httplib2==0.12.0
oauth2client==4.1.3
basicauth==0.4.1
python-dateutil==2.8.0